Confirmação antes de deletar. Mostra detalhes.

+++ semana7/labenu-users/labenu-users-system/src/components/UserList.js
@@ -4,7 +4,10 @@ import axios from 'axios'
 class UserList extends React.Component {
 
     state = {
-        usersList: []
+        usersList: [],
+        userDetail: false,
+        userName: "",
+        userId: ""
     }
 
     pullUserList = () => {
@@ -32,30 +35,47 @@ class UserList extends React.Component {
       }
     
       deleteUser = (id) => {
-        const request = axios.delete(`https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users/${id}`, 
-        {
-          headers: {
-            Authorization: "giovanna-caivano-jackson"
-          }
-        });
-    
-        request.then((answer) => {
-          alert("Usuário removido com sucesso")
-          this.pullUserList()
-        }).catch((error) => {
-          alert("Algo deu errado. Tente novamente.")
-        })
+
+        const confirm = prompt("Deseja mesmo remover este usuário? S/N").toLowerCase()
+        console.log(confirm)
+
+        if(confirm === "s") {
+            const request = axios.delete(`https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users/${id}`, 
+            {
+              headers: {
+                Authorization: "giovanna-caivano-jackson"
+              }
+            });
+        
+            request.then((answer) => {
+              alert("Usuário removido com sucesso")
+              this.pullUserList()
+            }).catch((error) => {
+              alert("Algo deu errado. Tente novamente.")
+            })
+        } else {
+            alert("Resposta inválida. Tente novamente.")
+        }
+
+      }
+
+      showDetail = (user) => {
+          this.setState({ userDetail: !this.state.userDetail, userName: user.name, userId: user.id })
+
       }
 
     render() {
         const renderedList = this.state.usersList.map((user) => {
-            return <li key={user.id}>{user.name} - <button onClick={() => this.deleteUser(user.id)}>X</button></li>
+            return <li key={user.id} onClick={() => this.showDetail(user)}>{user.name} - <button onClick={() => this.deleteUser(user.id)}>X</button></li>
         })
+
+        // const renderedUser = 
 
         return (
             <div>
                 <h1>Lista de Usuários Cadastrados</h1>
                 <ul>{renderedList}</ul>
+        <div> {this.state.userDetail && <p><b>Detalhes:</b> <br/> Nome: {this.state.userName} <br/> Email:</p>}</div>
             </div>
         )
     }
